Allow refreshing cached SMART attributes without reopening the disk
Users of this library can accomplish the same thing by reopening the disk, but
exposing refreshes like this is a little more convenient. Also, the existence of
this method makes it more obvious that data is cached.

The warning about wakeup from sleep is based on a warning in libatasmart.h.